Move payload functions into separate module
This change moves the helper functions that create the
`CircuitManagementPayload` objects that are used in the admin service
integration tests. This also alters the integration tests to use these
functions, imported from the `payload` module.
